Update fork from upstream.

+++ uninstall.php
@@ -4,7 +4,7 @@
  *
  * @package     EDD
  * @subpackage  Uninstall
- * @copyright   Copyright (c) 2013, Pippin Williamson
+ * @copyright   Copyright (c) 2014, Pippin Williamson
  * @license     http://opensource.org/licenses/gpl-2.0.php GNU Public License
  * @since       1.4.3
  */
@@ -17,53 +17,62 @@ include_once( 'easy-digital-downloads.php' );
 
 global $wpdb, $edd_options, $wp_roles;
 
-/** Delete All the Custom Post Types */
-$edd_post_types = array( 'download', 'edd_payment', 'edd_discount', 'edd_log' );
-foreach ( $edd_post_types as $post_type ) {
+if( edd_get_option( 'uninstall_on_delete' ) ) {
 
-	$items = get_posts( array( 'post_type' => $post_type, 'numberposts' => -1, 'fields' => 'ids' ) );
+	/** Delete All the Custom Post Types */
+	$edd_taxonomies = array( 'download_category', 'download_tag', 'edd_log_type', );
+	$edd_post_types = array( 'download', 'edd_payment', 'edd_discount', 'edd_log' );
+	foreach ( $edd_post_types as $post_type ) {
+	
+		$edd_taxonomies = array_merge( $edd_taxonomies, get_object_taxonomies( $post_type ) );
+		$items = get_posts( array( 'post_type' => $post_type, 'post_status' => 'any', 'numberposts' => -1, 'fields' => 'ids' ) );
 
-	if ( $items ) {
-		foreach ( $items as $item ) {
-			wp_delete_post( $item, true);
+		if ( $items ) {
+			foreach ( $items as $item ) {
+				wp_delete_post( $item, true);
+			}
 		}
 	}
-}
 
-/** Delete All the Taxonomies */
-$edd_taxonomies = array( 'download_tag', 'download_category', 'edd_log_type' );
-foreach ( $edd_taxonomies as $taxonomy ) {
-	global $wp_taxonomies;
-	$terms = get_terms( $taxonomy );
-	if ( $terms ) {
-		foreach ( $terms as $term ) {
-			wp_delete_term( $term->term_id, $taxonomy );
+	/** Delete All the Terms & Taxonomies */
+	foreach ( array_unique( array_filter( $edd_taxonomies ) ) as $taxonomy ) {
+		
+		$terms = $wpdb->get_results( $wpdb->prepare( "SELECT t.*, tt.* FROM $wpdb->terms AS t INNER JOIN $wpdb->term_taxonomy AS tt ON t.term_id = tt.term_id WHERE tt.taxonomy IN ('%s') ORDER BY t.name ASC", $taxonomy ) );
+		
+		// Delete Terms
+		if ( $terms ) {
+			foreach ( $terms as $term ) {
+				$wpdb->delete( $wpdb->term_taxonomy, array( 'term_taxonomy_id' => $term->term_taxonomy_id ) );
+				$wpdb->delete( $wpdb->terms, array( 'term_id' => $term->term_id ) );
+			}
+		}
+		
+		// Delete Taxonomies
+		$wpdb->delete( $wpdb->term_taxonomy, array( 'taxonomy' => $taxonomy ), array( '%s' ) );
+	}
+
+	/** Delete the Plugin Pages */
+	$edd_created_pages = array( 'purchase_page', 'success_page', 'failure_page', 'purchase_history_page' );
+	foreach ( $edd_created_pages as $p ) {
+		if ( isset( $edd_options[ $p ] ) ) {
+			wp_delete_post( $edd_options[ $p ], true );
 		}
 	}
-	unset( $wp_taxonomies[ $taxonomy ] );
-}
 
-/** Delete the Plugin Pages */
-if ( isset( $edd_options['purchase_page'] ) )
-	wp_delete_post( $edd_options['purchase_page'], true );
-if ( isset( $edd_options['success_page'] ) )
-	wp_delete_post( $edd_options['success_page'], true );
-if ( isset( $edd_options['failure_page'] ) )
-	wp_delete_post( $edd_options['failure_page'], true );
+	/** Delete all the Plugin Options */
+	delete_option( 'edd_settings' );
 
-/** Delete all the Plugin Options */
-delete_option( 'edd_settings_general' );
-delete_option( 'edd_settings_gateways' );
-delete_option( 'edd_settings_emails' );
-delete_option( 'edd_settings_styles' );
-delete_option( 'edd_settings_taxes' );
-delete_option( 'edd_settings_misc' );
+	/** Delete Capabilities */
+	EDD()->roles->remove_caps();
 
-/** Delete Capabilities */
-EDD()->roles->remove_caps();
+	/** Delete the Roles */
+	$edd_roles = array( 'shop_manager', 'shop_accountant', 'shop_worker', 'shop_vendor' );
+	foreach ( $edd_roles as $role ) {
+		remove_role( $role );
+	}
 
-/** Delete the Roles */
-$edd_roles = array( 'shop_manager', 'shop_accountant', 'shop_worker', 'shop_vendor' );
-foreach ( $edd_roles as $role ) {
-	remove_role( $role );
+	/** Cleanup Cron Events */
+	wp_clear_scheduled_hook( 'edd_daily_scheduled_events' );
+	wp_clear_scheduled_hook( 'edd_daily_cron' );
+	wp_clear_scheduled_hook( 'edd_weekly_cron' );
 }
